munchies/backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
argilla==1.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
autobahn==23.1.2
Automat==22.10.0
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.1.0
bleach==6.1.0
blinker==1.6.2
blis==0.7.11
Brotli==1.0.9
bs4==0.0.1
build==1.0.3
CacheControl==0.13.1
catalogue==2.0.10
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
cleo==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpathlib==0.16.0
comm==0.2.1
commonmark==0.9.1
confection==0.1.3
ConfigArgParse==1.5.3
constantly==15.1.0
contourpy==1.0.7
crashtest==0.4.1
cron-descriptor==1.2.35
cryptography==40.0.1
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.8
daphne==4.0.0
dataclasses-json==0.5.7
datasets==2.12.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.6
distlib==0.3.7
dj-database-url==1.2.0
Django==4.2.6
django-celery-beat==2.5.0
django-configurations==2.4.1
django-cors-headers==4.1.0
django-email-verification==0.3.1
django-environ==0.10.0
django-filter==23.1
django-nose==1.4.7
django-redis==5.2.0
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
dnspython==2.3.0
drf-spectacular==0.26.3
dulwich==0.21.6
email-validator==2.0.0.post2
environ==1.0
et-xmlfile==1.1.0
executing==1.2.0
Faker==18.4.0
fastapi==0.95.2
fastjsonschema==2.19.0
ffmpeg==1.4
ffprobe==0.5
filelock==3.13.1
Flask==2.3.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
gevent==22.10.2
geventhttpclient==2.0.9
googlemaps==4.10.0
greenlet==2.0.2
gTTS==2.4.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.17.3
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
inflection==0.5.1
installer==0.7.0
ipykernel==6.21.2
ipython==8.10.0
ipywidgets==8.1.2
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
keyring==24.3.0
kiwisolver==1.4.4
kombu==5.3.2
langchain==0.0.180
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
load-dotenv==0.1.0
locust==2.15.1
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
noisereduce==3.0.0
nose==1.3.7
notebook==7.0.8
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.4
numpy==1.24.3
olefile==0.46
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.8.14
overrides==7.7.0
packaging==23.1
pandas==1.5.3
pandoc==2.3
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.3
pathspec==0.11.1
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
pkginfo==1.9.6
platformdirs==3.11.0
plumbum==1.8.1
ply==3.11
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
pooch==1.8.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.38
Protego==0.3.0
protobuf==4.25.0
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.8
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.6.0
pylab-sdk==1.3.2
pymongo==4.6.0
pyOpenSSL==23.1.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.8.1
PyPDF2==3.0.1
pyproject_hooks==1.0.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.2
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
rapidfuzz==3.5.2
redis==4.5.3
referencing==0.29.1
regex==2023.5.5
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rouge==1.0.1
roundrobin==0.0.4
rpds-py==0.9.2
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.10.1
Scrapy==2.11.0
Send2Trash==1.8.2
sentencepiece==0.1.99
service-identity==21.1.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spark-submit==1.4.0
SQLAlchemy==2.0.15
sqlparse==0.4.3
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
terminado==0.18.0
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tldextract==5.0.1
tokenizers==0.14.1
tomlkit==0.12.3
torch==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
transformers==4.35.0
trove-classifiers==2023.11.14
Twisted==22.10.0
txaio==23.1.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.2
ujson==5.7.0
unstructured==0.6.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
virtualenv==20.24.6
w3lib==2.1.2
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==2.3.0
widgetsnbextension==4.0.10
wrapt==1.14.1
xattr==0.10.1
XlsxWriter==3.1.0
xxhash==3.2.0
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.17.0
zope.event==4.6
zope.interface==6.0
